refact: move GameObject construction to separate factory class & methods
